moved meszgen to myszka lib
